+++ test/main.test.ts
@@ -23,7 +23,7 @@ jest.mock("../src/inputs", () => inputs);
 jest.mock("../src/svgo", () => svgo);
 jest.mock("../src/templating", () => templating);
 
-import { PR_NOT_FOUND } from "../src/github-api";
+import { PR_NOT_FOUND } from "../src/constants";
 import main from "../src/main";
 
 
@@ -156,7 +156,7 @@ describe("Configuration", () => {
     "This should be a commit title",
     "Why not Zoidberg",
     "A templated commit title? {{optimizedCount}}",
-  ])("custom commit title (%s)", async (commitTitle) => {
+  ])("custom commit message title (%s)", async (commitTitle) => {
     const actionConfig = new inputs.ActionConfig();
     actionConfig.commitTitle = commitTitle;
 
@@ -173,12 +173,12 @@ describe("Configuration", () => {
   });
 
   test.each([
-    "This should be a commit desciption",
+    "This should be a commit body",
     "Shut up and take my money",
     "A templated commit title? {{filesList}}",
-  ])("custom commit description (%s)", async (commitDescription) => {
+  ])("custom commit message body (%s)", async (commitBoy) => {
     const actionConfig = new inputs.ActionConfig();
-    actionConfig.commitDescription = commitDescription;
+    actionConfig.commitBody = commitBoy;
 
     githubAPI.getPrNumber.mockReturnValueOnce(PR_NUMBER.ADD_SVG);
     inputs.ActionConfig.mockReturnValueOnce(actionConfig);
@@ -187,7 +187,7 @@ describe("Configuration", () => {
 
     expect(templating.formatCommitMessage).toHaveBeenCalledWith(
       expect.any(String),
-      commitDescription,
+      commitBoy,
       expect.any(Object),
     );
   });